✨ Added lyrics support
ℹ️ Implemented functionality to fetch and display song lyrics within the application. This feature includes:
- Integration with a lyrics API to retrieve lyrics for a given song.
- UI updates to dynamically display lyrics alongside the song playback.

+++ lib/screens/home_screen/section_item.dart
@@ -336,7 +336,7 @@ class _ItemListState extends State<ItemList> {
                           backgroundImage: CachedNetworkImageProvider(
                             getEnhancedImage(
                                 items[index]['thumbnails'].first['url'],
-                                width: width),
+                                width: width.toInt()),
                           ),
                           radius: height / 2,
                         )
@@ -345,7 +345,7 @@ class _ItemListState extends State<ItemList> {
                           child: CachedNetworkImage(
                             imageUrl: getEnhancedImage(
                               items[index]['thumbnails'].first['url'],
-                              width: width,
+                              width: width.toInt(),
                             ),
                             height: height,
                             width: width,
